% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_ggstrusi.R
\docType{package}
\name{ggstrusi}
\alias{ggstrusi}
\title{Some functions to Be used with 'ggplot2'}
\description{
Some functions to be used with 'ggplot2'. One example is the creation of a 
version of the boxplot proposed by Edward Tufte.
}
