% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tufte_boxplot.R
\name{prepare_tufte}
\alias{prepare_tufte}
\title{Preparation of a point and two lineranges for a boxplot variant of Tufte}
\usage{
prepare_tufte(x = "", ymin, lower, middle, upper, ymax, ...)
}
\arguments{
\item{x}{character string with the name of the variable to be plotted on the x-axis}

\item{ymin}{name of the column with the minimum value of y to be plotted}

\item{lower}{name of the column with the lower part of the "box" of the boxplot}

\item{middle}{name of the column with the middle point of the boxplot (usually the median)}

\item{upper}{name of the column with the upper part of the "box" of the boxplot}

\item{ymax}{name of the column with the maximum value of y to be plotted}

\item{...}{other aesthetic mappings created by \code{\link[ggplot2]{aes_string}}, passed to the 
\code{\link[ggplot2]{geom_point}} element.}
}
\value{
a list with three elements: two \code{\link[ggplot2]{geom_linerange}}s and a \code{\link[ggplot2]{geom_point}}.
}
\description{
This function prepares the elements necessary to draw a variant of the boxplot
proposed by Edward Tufte.
}
\examples{
data("c02_14")
l <- prepare_tufte(x = "sesso", ymin = "p10", lower = "p25",
                              middle = "p50", upper = "p75", ymax = "p90")
                              
# Tufte boxplot of the first row of the data:
tufte_boxplot(data = c02_14[1, ], l)

# Tufte boxplot of the first three rows:
tufte_boxplot(data = c02_14[1:3, ], l)

l
str(l)

# An example where we plot boxplots for each group:
l <- prepare_tufte(x = "id", ymin = "p10", lower = "p25",
                              middle = "p50", upper = "p75", ymax = "p90")
tufte_boxplot(data = c02_14[1, ], l)
}
\seealso{
\code{\link{prepare_tufte_}} which does quite the same things but, instead of character arguments, uses 
one-sided \code{\link[stats]{formula}} to set aesthetic mappings.
}
